+++ Gruntfile.js
@@ -1,23 +1,8 @@
 module.exports = function (grunt) {
 	grunt.initConfig({
 		pkg : grunt.file.readJSON("package.json"),
-		babel: {
-			options: {
-				sourceMap: false,
-				presets: ["babel-preset-es2015"]
-			},
-			dist: {
-				files: [{
-					expand: true,
-					cwd: 'src',
-					src: ['*.js'],
-					dest: 'lib',
-					ext: '.js'
-				}]
-			}
-		},
 		eslint: {
-			target: ["src/*.js"]
+			target: ["lib/*.js"]
 		},
 		mochaTest : {
 			options: {
@@ -68,7 +53,6 @@ module.exports = function (grunt) {
 	});
 
 	// tasks
-	grunt.loadNpmTasks("grunt-babel");
 	grunt.loadNpmTasks("grunt-eslint");
 	grunt.loadNpmTasks("grunt-mocha-test");
 	grunt.loadNpmTasks("grunt-nsp");
@@ -78,6 +62,6 @@ module.exports = function (grunt) {
 
 	// aliases
 	grunt.registerTask("test", ["eslint", "mochaTest", "nsp"]);
-	grunt.registerTask("build", ["sass", "babel", "sed"]);
+	grunt.registerTask("build", ["sass", "sed"]);
 	grunt.registerTask("default", ["build", "test"]);
 };
