+++ lib/iterate.js
@@ -1,6 +1,4 @@
-"use strict";
-
-function iterate(obj, fn) {
+function iterate (obj, fn) {
 	if (obj instanceof Object) {
 		Object.keys(obj).forEach(function (i) {
 			fn.call(obj, obj[i], i);
