+++ lib/middleware.js
@@ -1,14 +1,16 @@
-"use strict";
+const path = require("path"),
+	array = require("retsu"),
+	coerce = require("tiny-coerce"),
+	regex = require(path.join(__dirname, "regex.js")),
+	iterate = require(path.join(__dirname, "iterate.js"));
 
-var path = require("path"),
-    array = require("retsu"),
-    coerce = require("tiny-coerce"),
-    regex = require(path.join(__dirname, "regex.js")),
-    iterate = require(path.join(__dirname, "iterate.js"));
+const rateHeaders = [
+	"x-ratelimit-limit",
+	"x-ratelimit-remaining",
+	"x-ratelimit-reset"
+];
 
-var rateHeaders = ["x-ratelimit-limit", "x-ratelimit-remaining", "x-ratelimit-reset"];
-
-function decorate(req, res, next) {
+function decorate (req, res, next) {
 	var obj = req.server.tenso;
 
 	req.protect = false;
@@ -19,9 +21,7 @@ function decorate(req, res, next) {
 		return obj.error(req, res, status, body);
 	};
 
-	res.redirect = function (uri) {
-		var perm = arguments.length <= 1 || arguments[1] === undefined ? true : arguments[1];
-
+	res.redirect = function (uri, perm = true) {
 		return obj.redirect(req, res, uri, undefined, perm);
 	};
 
@@ -36,15 +36,15 @@ function decorate(req, res, next) {
 	next();
 }
 
-function asyncFlag(req, res, next) {
+function asyncFlag (req, res, next) {
 	req.protectAsync = true;
 	next();
 }
 
-function bypass(req, res, next) {
-	var pass = req.server.config.auth.unprotect.filter(function (i) {
-		return i.test(req.url);
-	}).length > 0;
+function bypass (req, res, next) {
+	let pass = req.server.config.auth.unprotect.filter(function (i) {
+			return i.test(req.url);
+		}).length > 0;
 
 	if (pass) {
 		req.unprotect = true;
@@ -53,7 +53,7 @@ function bypass(req, res, next) {
 	next();
 }
 
-function guard(req, res, next) {
+function guard (req, res, next) {
 	if (req.parsed.url === "/login" || req.isAuthenticated()) {
 		next();
 	} else {
@@ -61,9 +61,8 @@ function guard(req, res, next) {
 	}
 }
 
-function parse(req, res, next) {
-	var args = void 0,
-	    type = void 0;
+function parse (req, res, next) {
+	let args, type;
 
 	if (regex.body.test(req.method) && req.body !== undefined) {
 		type = req.headers["content-type"];
@@ -89,13 +88,10 @@ function parse(req, res, next) {
 	next();
 }
 
-function keymaster(req, res, next) {
-	var obj = req.server.tenso,
-	    authd = req.session && req.isAuthenticated(),
-	    method = void 0,
-	    result = void 0,
-	    routes = void 0,
-	    uri = void 0;
+function keymaster (req, res, next) {
+	let obj = req.server.tenso,
+		authd = req.session && req.isAuthenticated(),
+		method, result, routes, uri;
 
 	// No authentication, or it's already happened
 	if (!req.protect || !req.protectAsync || authd) {
@@ -133,11 +129,9 @@ function keymaster(req, res, next) {
 	}
 }
 
-function rate(req, res, next) {
-	var config = void 0,
-	    good = void 0,
-	    results = void 0,
-	    server = void 0;
+
+function rate (req, res, next) {
+	let config, good, results, server;
 
 	if (req.unprotect) {
 		next();
@@ -159,9 +153,9 @@ function rate(req, res, next) {
 	}
 }
 
-function zuul(req, res, next) {
-	var uri = req.parsed.path,
-	    protectd = false;
+function zuul (req, res, next) {
+	let uri = req.parsed.path,
+		protectd = false;
 
 	array.each(req.server.config.auth.protect, function (r) {
 		if (r.test(uri)) {
@@ -184,7 +178,7 @@ function zuul(req, res, next) {
 	});
 }
 
-function valid(req, res, next) {
+function valid (req, res, next) {
 	if (req.allow.indexOf(req.method) > -1) {
 		next();
 	} else {
