+++ lib/regex.js
@@ -1,6 +1,4 @@
-"use strict";
-
-var regex = {
+const regex = {
 	body: /POST|PUT|PATCH/i,
 	body_split: /&|=/,
 	collection: /(.*)(\/.*)$/,
