+++ lib/renderers.js
@@ -1,16 +1,14 @@
-"use strict";
+const array = require("retsu"),
+	xml = require("tiny-xml"),
+	yaml = require("yamljs"),
+	path = require("path"),
+	csv = require("csv.js"),
+	utility = require(path.join(__dirname, "utility.js"));
 
-var array = require("retsu"),
-    xml = require("tiny-xml"),
-    yaml = require("yamljs"),
-    path = require("path"),
-    csv = require("csv.js"),
-    utility = require(path.join(__dirname, "utility.js"));
+let renderers = new Map();
 
-var renderers = new Map();
-
-function sanitize(arg) {
-	var output = arg;
+function sanitize (arg) {
+	let output = arg;
 
 	if (typeof arg === "string") {
 		[["<", "&lt;"], [">", "&gt;"]].forEach(function (i) {
@@ -30,19 +28,27 @@ renderers.set("text/csv", function (arg, req, headers) {
 });
 
 renderers.set("text/html", function (arg, req, headers, tpl) {
-	var protocol = req.headers["x-forwarded-proto"] ? req.headers["x-forwarded-proto"] + ":" : req.parsed.protocol;
+	let protocol = req.headers["x-forwarded-proto"] ? req.headers["x-forwarded-proto"] + ":" : req.parsed.protocol;
 
-	return (tpl || "").replace(/\{\{title\}\}/g, req.server.config.title).replace("{{url}}", req.parsed.href.replace(req.parsed.protocol, protocol)).replace("{{headers}}", Object.keys(headers).sort(array.sort).map(function (i) {
-		return "<tr><td>" + i + "</td><td>" + sanitize(headers[i]) + "</td></tr>";
-	}).join("\n")).replace("{{formats}}", "<option value=''></option>" + Array.from(renderers.keys()).filter(function (i) {
-		return i.indexOf("html") === -1;
-	}).map(function (i) {
-		return "<option value='" + i + "'>" + i.replace(/^.*\//, "").toUpperCase() + "</option>";
-	}).join("\n")).replace("{{body}}", JSON.stringify(arg, null, 2)).replace("{{year}}", new Date().getFullYear()).replace("{{version}}", "3.3.13").replace("{{allow}}", headers.allow).replace("{{methods}}", utility.explode(headers.allow.replace("GET, HEAD, OPTIONS", "")).filter(function (i) {
-		return i !== "";
-	}).map(function (i) {
-		return "<option value='" + i + "'>" + i + "</option>";
-	}).join("\n")).replace("{{csrf}}", headers["x-csrf-token"] || "");
+	return (tpl || "")
+		.replace(/\{\{title\}\}/g, req.server.config.title)
+		.replace("{{url}}", req.parsed.href.replace(req.parsed.protocol, protocol))
+		.replace("{{headers}}", Object.keys(headers).sort(array.sort).map(function (i) {
+			return "<tr><td>" + i + "</td><td>" + sanitize(headers[i]) + "</td></tr>";
+		}).join("\n"))
+		.replace("{{formats}}", "<option value=''></option>" + Array.from(renderers.keys()).filter(function (i) { return i.indexOf("html") === -1; }).map(function (i) {
+			return "<option value='" + i + "'>" + i.replace(/^.*\//, "").toUpperCase() + "</option>";
+		}).join("\n"))
+		.replace("{{body}}", JSON.stringify(arg, null, 2))
+		.replace("{{year}}", new Date().getFullYear())
+		.replace("{{version}}", "4.0.0")
+		.replace("{{allow}}", headers.allow)
+		.replace("{{methods}}", utility.explode(headers.allow.replace("GET, HEAD, OPTIONS", "")).filter(function (i) {
+			return i !== "";
+		}).map(function (i) {
+			return "<option value='" + i + "'>" + i + "</option>";
+		}).join("\n"))
+		.replace("{{csrf}}", headers["x-csrf-token"] || "");
 });
 
 renderers.set("application/javascript", function (arg, req, headers) {
