+++ lib/serializers.js
@@ -1,9 +1,7 @@
-"use strict";
-
-var serializers = new Map();
+let serializers = new Map();
 
 /** Default serializer */
-function tenso(arg, err, status) {
+function tenso (arg, err, status) {
 	return {
 		data: arg !== null ? arg : null,
 		error: arg === null ? err.message || err || "Something went wrong" : null,
