+++ lib/utility.js
@@ -1,58 +1,51 @@
-"use strict";
+const path = require("path"),
+	http = require("http"),
+	array = require("retsu"),
+	coerce = require("tiny-coerce"),
+	keysort = require("keysort"),
+	url = require("url"),
+	session = require("express-session"),
+	cookie = require("cookie-parser"),
+	lusca = require("lusca"),
+	uuid = require("tiny-uuid4"),
+	middleware = require(path.join(__dirname, "middleware.js")),
+	regex = require(path.join(__dirname, "regex.js")),
+	iterate = require(path.join(__dirname, "iterate.js")),
+	passport = require("passport"),
+	BasicStrategy = require("passport-http").BasicStrategy,
+	BearerStrategy = require("passport-http-bearer").Strategy,
+	FacebookStrategy = require("passport-facebook").Strategy,
+	GoogleStrategy = require("passport-google").Strategy,
+	LinkedInStrategy = require("passport-linkedin").Strategy,
+	LocalStrategy = require("passport-local").Strategy,
+	OAuth2Strategy = require("passport-oauth2").Strategy,
+	SAMLStrategy = require("passport-saml").Strategy,
+	TwitterStrategy = require("passport-twitter").Strategy,
+	RedisStore = require("connect-redis")(session);
 
-var path = require("path"),
-    http = require("http"),
-    array = require("retsu"),
-    coerce = require("tiny-coerce"),
-    keysort = require("keysort"),
-    url = require("url"),
-    session = require("express-session"),
-    cookie = require("cookie-parser"),
-    lusca = require("lusca"),
-    uuid = require("tiny-uuid4"),
-    middleware = require(path.join(__dirname, "middleware.js")),
-    regex = require(path.join(__dirname, "regex.js")),
-    iterate = require(path.join(__dirname, "iterate.js")),
-    passport = require("passport"),
-    BasicStrategy = require("passport-http").BasicStrategy,
-    BearerStrategy = require("passport-http-bearer").Strategy,
-    FacebookStrategy = require("passport-facebook").Strategy,
-    GoogleStrategy = require("passport-google").Strategy,
-    LinkedInStrategy = require("passport-linkedin").Strategy,
-    LocalStrategy = require("passport-local").Strategy,
-    OAuth2Strategy = require("passport-oauth2").Strategy,
-    SAMLStrategy = require("passport-saml").Strategy,
-    TwitterStrategy = require("passport-twitter").Strategy,
-    RedisStore = require("connect-redis")(session);
-
-var lws = void 0,
-    coap = void 0;
+let uws, coap;
 
 try {
-	lws = require("lws");
+	uws = require("uws");
 	coap = require("coap");
 } catch (e) {
 	void 0;
 }
 
-function trim(obj) {
+function trim (obj) {
 	return obj.replace(/^(\s+|\t+|\n+)|(\s+|\t+|\n+)$/g, "");
 }
 
-function explode(obj) {
-	var arg = arguments.length <= 1 || arguments[1] === undefined ? "," : arguments[1];
-
+function explode (obj, arg = ",") {
 	return trim(obj).split(new RegExp("\\s*" + arg + "\\s*"));
 }
 
-function escape(arg) {
+function escape (arg) {
 	return arg.replace(/[-[\]{}()*+?.,\\^$|#]/g, "\\$&");
 }
 
-function capitalize(obj) {
-	var all = arguments.length <= 1 || arguments[1] === undefined ? false : arguments[1];
-
-	var result = void 0;
+function capitalize (obj, all = false) {
+	let result;
 
 	if (all) {
 		result = explode(obj, " ").map(capitalize).join(" ");
@@ -63,38 +56,27 @@ function capitalize(obj) {
 	return result;
 }
 
-function clone(arg) {
+function clone (arg) {
 	return JSON.parse(JSON.stringify(arg));
 }
 
-function isEmpty(obj) {
+function isEmpty (obj) {
 	return trim(obj) === "";
 }
 
-function auth(obj, config) {
-	var ssl = config.ssl.cert && config.ssl.key,
-	    proto = "http" + (ssl ? "s" : ""),
-	    realm = proto + "://" + (config.hostname === "localhost" ? "127.0.0.1" : config.hostname) + (config.port !== 80 && config.port !== 443 ? ":" + config.port : ""),
-	    async = (config.auth.facebook.enabled || config.auth.google.enabled || config.auth.linkedin.enabled || config.auth.twitter.enabled) !== false,
-	    stateless = (config.auth.basic.enabled || config.auth.bearer.enabled) !== false,
-	    stateful = (async || config.auth.local.enabled || config.security.csrf) !== false,
-	    authMap = {},
-	    authUris = [],
-	    keys = void 0,
-	    sesh = void 0,
-	    fnCookie = void 0,
-	    fnSession = void 0,
-	    luscaCsp = void 0,
-	    luscaCsrf = void 0,
-	    luscaXframe = void 0,
-	    luscaP3p = void 0,
-	    luscaHsts = void 0,
-	    luscaXssProtection = void 0,
-	    passportAuth = void 0,
-	    passportInit = void 0,
-	    passportSession = void 0;
+function auth (obj, config) {
+	let ssl = config.ssl.cert && config.ssl.key,
+		proto = "http" + (ssl ? "s" : ""),
+		realm = proto + "://" + (config.hostname === "localhost" ? "127.0.0.1" : config.hostname) + (config.port !== 80 && config.port !== 443 ? ":" + config.port : ""),
+		async = (config.auth.facebook.enabled || config.auth.google.enabled || config.auth.linkedin.enabled || config.auth.twitter.enabled) !== false,
+		stateless = (config.auth.basic.enabled || config.auth.bearer.enabled) !== false,
+		stateful = (async || config.auth.local.enabled || config.security.csrf) !== false,
+		authMap = {},
+		authUris = [],
+		keys, sesh, fnCookie, fnSession, luscaCsp, luscaCsrf, luscaXframe, luscaP3p, luscaHsts, luscaXssProtection, passportAuth,
+		passportInit, passportSession;
 
-	function csrfWrapper(req, res, next) {
+	function csrfWrapper (req, res, next) {
 		if (req.unprotect) {
 			next();
 		} else {
@@ -102,7 +84,7 @@ function auth(obj, config) {
 		}
 	}
 
-	function init(sess) {
+	function init (sess) {
 		passportInit = passport.initialize();
 		obj.server.use(passportInit).blacklist(passportInit);
 
@@ -112,7 +94,7 @@ function auth(obj, config) {
 		}
 	}
 
-	function redirect(req, res) {
+	function redirect (req, res) {
 		res.redirect(config.auth.redirect);
 	}
 
@@ -167,7 +149,7 @@ function auth(obj, config) {
 		obj.server.use(middleware.bypass).blacklist(middleware.bypass);
 
 		if (config.security.csrf) {
-			luscaCsrf = lusca.csrf({ key: config.security.key, secret: config.security.secret });
+			luscaCsrf = lusca.csrf({key: config.security.key, secret: config.security.secret});
 			obj.server.use(csrfWrapper).blacklist(csrfWrapper);
 		}
 	}
@@ -221,7 +203,7 @@ function auth(obj, config) {
 			}
 
 			(function () {
-				var r = "(?!/auth/(";
+				let r = "(?!/auth/(";
 
 				array.each(authUris, function (i) {
 					r += i.replace("_uri", "") + "|";
@@ -229,14 +211,14 @@ function auth(obj, config) {
 
 				r = r.replace(/\|$/, "") + ")).*$";
 				obj.server.use(r, middleware.guard).blacklist(middleware.guard);
-			})();
+			}());
 
 			config.routes.get["/login"] = config.auth.local.enabled ? keys ? {
 				login_uri: "/auth",
 				instruction: "POST 'username' & 'password' to authenticate"
-			} : { instruction: "POST 'username' & 'password' to authenticate" } : { login_uri: "/auth" };
+			} : {instruction: "POST 'username' & 'password' to authenticate"} : {login_uri: "/auth"};
 		} else if (config.auth.local.enabled) {
-			config.routes.get["/login"] = { instruction: "POST 'username' & 'password' to authenticate" };
+			config.routes.get["/login"] = {instruction: "POST 'username' & 'password' to authenticate"};
 		}
 
 		config.routes.get["/logout"] = function (req, res) {
@@ -250,9 +232,9 @@ function auth(obj, config) {
 
 	if (config.auth.basic.enabled) {
 		(function () {
-			var x = {};
+			let x = {};
 
-			function validate(arg, cb) {
+			function validate (arg, cb) {
 				if (x[arg]) {
 					cb(null, x[arg]);
 				} else {
@@ -261,10 +243,10 @@ function auth(obj, config) {
 			}
 
 			array.each(config.auth.basic.list || [], function (i) {
-				var args = i.split(":");
+				let args = i.split(":");
 
 				if (args.length > 0) {
-					x[args[0]] = { password: args[1] };
+					x[args[0]] = {password: args[1]};
 				}
 			});
 
@@ -283,7 +265,7 @@ function auth(obj, config) {
 				});
 			}));
 
-			passportAuth = passport.authenticate("basic", { session: stateful });
+			passportAuth = passport.authenticate("basic", {session: stateful});
 
 			if (async || config.auth.local.enabled) {
 				obj.server.get("/auth/basic", passportAuth).blacklist(passportAuth);
@@ -291,14 +273,14 @@ function auth(obj, config) {
 			} else {
 				obj.server.use(passportAuth).blacklist(passportAuth);
 			}
-		})();
+		}());
 	}
 
 	if (config.auth.bearer.enabled) {
 		(function () {
-			var x = config.auth.bearer.tokens || [];
+			let x = config.auth.bearer.tokens || [];
 
-			function validate(arg, cb) {
+			function validate (arg, cb) {
 				if (array.contains(x, arg)) {
 					cb(null, arg);
 				} else {
@@ -317,11 +299,11 @@ function auth(obj, config) {
 						return done(null, false);
 					}
 
-					return done(null, user, { scope: "read" });
+					return done(null, user, {scope: "read"});
 				});
 			}));
 
-			passportAuth = passport.authenticate("bearer", { session: stateful });
+			passportAuth = passport.authenticate("bearer", {session: stateful});
 
 			if (async || config.auth.local.enabled) {
 				obj.server.get("/auth/bearer", passportAuth).blacklist(passportAuth);
@@ -329,7 +311,7 @@ function auth(obj, config) {
 			} else {
 				obj.server.use(passportAuth).blacklist(passportAuth);
 			}
-		})();
+		}());
 	}
 
 	if (config.auth.facebook.enabled) {
@@ -351,7 +333,7 @@ function auth(obj, config) {
 		obj.server.get("/auth/facebook", middleware.asyncFlag);
 		obj.server.get("/auth/facebook", passport.authenticate("facebook"));
 		obj.server.get("/auth/facebook/callback", middleware.asyncFlag);
-		obj.server.get("/auth/facebook/callback", passport.authenticate("facebook", { failureRedirect: "/login" }));
+		obj.server.get("/auth/facebook/callback", passport.authenticate("facebook", {failureRedirect: "/login"}));
 		obj.server.get("/auth/facebook/callback", redirect);
 	}
 
@@ -373,7 +355,7 @@ function auth(obj, config) {
 		obj.server.get("/auth/google", middleware.asyncFlag);
 		obj.server.get("/auth/google", passport.authenticate("google"));
 		obj.server.get("/auth/google/callback", middleware.asyncFlag);
-		obj.server.get("/auth/google/callback", passport.authenticate("google", { failureRedirect: "/login" }));
+		obj.server.get("/auth/google/callback", passport.authenticate("google", {failureRedirect: "/login"}));
 		obj.server.get("/auth/google/callback", redirect);
 	}
 
@@ -394,9 +376,9 @@ function auth(obj, config) {
 		}));
 
 		obj.server.get("/auth/linkedin", middleware.asyncFlag);
-		obj.server.get("/auth/linkedin", passport.authenticate("linkedin", { "scope": config.auth.linkedin.scope || ["r_basicprofile", "r_emailaddress"] }));
+		obj.server.get("/auth/linkedin", passport.authenticate("linkedin", {"scope": config.auth.linkedin.scope || ["r_basicprofile", "r_emailaddress"]}));
 		obj.server.get("/auth/linkedin/callback", middleware.asyncFlag);
-		obj.server.get("/auth/linkedin/callback", passport.authenticate("linkedin", { failureRedirect: "/login" }));
+		obj.server.get("/auth/linkedin/callback", passport.authenticate("linkedin", {failureRedirect: "/login"}));
 		obj.server.get("/auth/linkedin/callback", redirect);
 	}
 
@@ -414,7 +396,7 @@ function auth(obj, config) {
 
 		config.routes.post = config.routes.post || {};
 		config.routes.post["/login"] = function (req, res) {
-			function final() {
+			function final () {
 				passport.authenticate("local")(req, res, function (e) {
 					if (e) {
 						res.error(401, "Unauthorized");
@@ -426,7 +408,7 @@ function auth(obj, config) {
 				});
 			}
 
-			function mid() {
+			function mid () {
 				passportSession(req, res, final);
 			}
 
@@ -455,13 +437,13 @@ function auth(obj, config) {
 		obj.server.get("/auth/oauth2", middleware.asyncFlag);
 		obj.server.get("/auth/oauth2", passport.authenticate("oauth2"));
 		obj.server.get("/auth/oauth2/callback", middleware.asyncFlag);
-		obj.server.get("/auth/oauth2/callback", passport.authenticate("oauth2", { failureRedirect: "/login" }));
+		obj.server.get("/auth/oauth2/callback", passport.authenticate("oauth2", {failureRedirect: "/login"}));
 		obj.server.get("/auth/oauth2/callback", redirect);
 	}
 
 	if (config.auth.saml.enabled) {
 		(function () {
-			var arg = config.auth.saml;
+			let arg = config.auth.saml;
 
 			arg.callbackURL = realm + "/auth/saml/callback";
 			delete arg.enabled;
@@ -477,12 +459,12 @@ function auth(obj, config) {
 					done(null, user);
 				});
 			}));
-		})();
+		}());
 
 		obj.server.get("/auth/saml", middleware.asyncFlag);
 		obj.server.get("/auth/saml", passport.authenticate("saml"));
 		obj.server.get("/auth/saml/callback", middleware.asyncFlag);
-		obj.server.get("/auth/saml/callback", passport.authenticate("saml", { failureRedirect: "/login" }));
+		obj.server.get("/auth/saml/callback", passport.authenticate("saml", {failureRedirect: "/login"}));
 		obj.server.get("/auth/saml/callback", redirect);
 	}
 
@@ -514,7 +496,7 @@ function auth(obj, config) {
 	return config;
 }
 
-function bootstrap(obj, config) {
+function bootstrap (obj, config) {
 	// Bootstrapping configuration
 	auth(obj, config);
 
@@ -523,17 +505,17 @@ function bootstrap(obj, config) {
 		config.headers = {};
 	}
 
-	config.headers.server = "tenso/3.3.13";
+	config.headers.server = "tenso/4.0.0";
 
 	// Starting WebSocket server
-	if (config.websocket.enabled && lws) {
-		obj.websocket = new lws.Server(config.websocket.options);
+	if (config.websocket.enabled && uws) {
+		obj.websocket = new uws.Server(config.websocket.options);
 		obj.server.log("Started WebSocket server on port " + config.websocket.options.port, "debug");
 	}
 
 	// Starting COAP server
 	if (config.coap.enabled && coap) {
-		obj.coap = coap.createServer({ type: config.coap.options.type });
+		obj.coap = coap.createServer({type: config.coap.options.type});
 		obj.coap.listen(config.coap.options.port, config.hostname);
 		obj.server.log("Started COAP (" + config.coap.options.type + ") server on " + config.hostname + ":" + config.coap.options.port, "debug");
 	}
@@ -544,7 +526,7 @@ function bootstrap(obj, config) {
 			if (obj.websocket) {
 				iterate(routes, function (fn, event) {
 					obj.server.log("COAP event handler: '" + event + "'", "debug");
-					obj.coap.on(event, function (req, res) {
+					obj.coap.on(event, (req, res) => {
 						fn(req, res, obj.coap, obj);
 					});
 				});
@@ -553,7 +535,7 @@ function bootstrap(obj, config) {
 			if (obj.websocket) {
 				iterate(routes, function (fn, event) {
 					obj.server.log("WebSocket event handler: '" + event + "'", "debug");
-					obj.websocket.on(event, function (socket, message, binary) {
+					obj.websocket.on(event, (socket, message, binary) => {
 						if (event === "message") {
 							fn(socket, message, binary, obj.websocket, obj);
 						} else {
@@ -565,11 +547,7 @@ function bootstrap(obj, config) {
 		} else {
 			iterate(routes, function (arg, route) {
 				if (typeof arg === "function") {
-					obj.server[method](route, function () {
-						for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-							args[_key] = arguments[_key];
-						}
-
+					obj.server[method](route, function (...args) {
 						arg.apply(obj, args);
 					});
 				} else {
@@ -591,8 +569,8 @@ function bootstrap(obj, config) {
 
 	// Starting API server
 	obj.server.start(config, function (req, res, status, msg) {
-		var stat = status instanceof Error ? parseInt(status.message, 10) : status,
-		    err = msg instanceof Error ? msg : new Error(msg || http.STATUS_CODES[stat]);
+		let stat = status instanceof Error ? parseInt(status.message, 10) : status,
+			err = msg instanceof Error ? msg : new Error(msg || http.STATUS_CODES[stat]);
 
 		obj.respond(req, res, err, status);
 	});
@@ -600,12 +578,10 @@ function bootstrap(obj, config) {
 	return obj;
 }
 
-function queryString() {
-	var qstring = arguments.length <= 0 || arguments[0] === undefined ? "" : arguments[0];
-
-	var obj = {};
-	var aresult = qstring.split("?");
-	var result = void 0;
+function queryString (qstring = "") {
+	let obj = {};
+	let aresult = qstring.split("?");
+	let result;
 
 	if (aresult.length > 1) {
 		aresult.shift();
@@ -613,8 +589,8 @@ function queryString() {
 
 	result = aresult.join("?");
 	result.split("&").forEach(function (prop) {
-		var aitem = prop.replace(/\+/g, " ").split("=");
-		var item = void 0;
+		let aitem = prop.replace(/\+/g, " ").split("=");
+		let item;
 
 		if (aitem.length > 2) {
 			item = [aitem.shift(), aitem.join("=")];
@@ -645,11 +621,9 @@ function queryString() {
 	return obj;
 }
 
-function parse(uri) {
-	var luri = uri;
-	var idxAscii = void 0,
-	    idxQ = void 0,
-	    parsed = void 0;
+function parse (uri) {
+	let luri = uri;
+	let idxAscii, idxQ, parsed;
 
 	if (luri === undefined || luri === null) {
 		luri = "";
@@ -679,29 +653,22 @@ function parse(uri) {
 	return parsed;
 }
 
-function hypermedia(server, req, rep, headers) {
-	var seen = {},
-	    collection = req.parsed.pathname,
-	    query = void 0,
-	    page = void 0,
-	    page_size = void 0,
-	    nth = void 0,
-	    root = void 0,
-	    proot = void 0,
-	    parent = void 0;
+function hypermedia (server, req, rep, headers) {
+	let seen = {},
+		collection = req.parsed.pathname,
+		query, page, page_size, nth, root, proot, parent;
 
 	// Parsing the object for hypermedia properties
-	function marshal(obj, rel, item_collection) {
-		var keys = Object.keys(obj),
-		    lrel = rel || "related",
-		    result = void 0;
+	function marshal (obj, rel, item_collection) {
+		let keys = Object.keys(obj),
+			lrel = rel || "related",
+			result;
 
 		if (keys.length === 0) {
 			result = null;
 		} else {
 			array.each(keys, function (i) {
-				var lcollection = void 0,
-				    uri = void 0;
+				let lcollection, uri;
 
 				// If ID like keys are found, and are not URIs, they are assumed to be root collections
 				if (regex.id.test(i) || regex.hypermedia.test(i)) {
@@ -719,7 +686,7 @@ function hypermedia(server, req, rep, headers) {
 						seen[uri] = 1;
 
 						if (server.allowed("GET", uri, req.host)) {
-							rep.links.push({ uri: uri, rel: lrel });
+							rep.links.push({uri: uri, rel: lrel});
 						}
 					}
 				}
@@ -762,40 +729,40 @@ function hypermedia(server, req, rep, headers) {
 					query.page_size = page_size;
 
 					root += "?" + Object.keys(query).map(function (i) {
-						return i + "=" + encodeURIComponent(query[i]);
-					}).join("&");
+							return i + "=" + encodeURIComponent(query[i]);
+						}).join("&");
 
 					if (page > 1) {
-						rep.links.push({ uri: root.replace("page=0", "page=1"), rel: "first" });
+						rep.links.push({uri: root.replace("page=0", "page=1"), rel: "first"});
 					}
 
 					if (page - 1 > 1 && page <= nth) {
-						rep.links.push({ uri: root.replace("page=0", "page=" + (page - 1)), rel: "prev" });
+						rep.links.push({uri: root.replace("page=0", "page=" + (page - 1)), rel: "prev"});
 					}
 
 					if (page + 1 < nth) {
-						rep.links.push({ uri: root.replace("page=0", "page=" + (page + 1)), rel: "next" });
+						rep.links.push({uri: root.replace("page=0", "page=" + (page + 1)), rel: "next"});
 					}
 
 					if (nth > 0 && page !== nth) {
-						rep.links.push({ uri: root.replace("page=0", "page=" + nth), rel: "last" });
+						rep.links.push({uri: root.replace("page=0", "page=" + nth), rel: "last"});
 					}
 				} else {
 					root += "?" + Object.keys(query).map(function (i) {
-						return i + "=" + encodeURIComponent(query[i]);
-					}).join("&");
+							return i + "=" + encodeURIComponent(query[i]);
+						}).join("&");
 				}
 			}
 
 			array.each(rep.data, function (i) {
 				var li = i.toString(),
-				    uri;
+					uri;
 
 				if (li !== collection) {
 					uri = li.indexOf("//") > -1 || li.indexOf("/") === 0 ? li : (collection + "/" + li).replace(/^\/\//, "/");
 
 					if (server.allowed("GET", uri, req.host)) {
-						rep.links.push({ uri: uri, rel: "item" });
+						rep.links.push({uri: uri, rel: "item"});
 					}
 				}
