--- src/iterate.js
@@ -1,11 +1,0 @@
-function iterate (obj, fn) {
-	if (obj instanceof Object) {
-		Object.keys(obj).forEach(function (i) {
-			fn.call(obj, obj[i], i);
-		});
-	} else {
-		obj.forEach(fn);
-	}
-}
-
-module.exports = iterate;
