+++ test/auth_test.js
@@ -1,6 +1,6 @@
 var hippie = require("hippie"),
 	emitter = require("events"),
-	tenso = require("../lib/index"),
+	tenso = require("../index"),
 	routes = require("./routes.js"),
 	array = require("retsu"),
 	csrf = 'x-csrf-token';
