--- www/css/style.css
@@ -1,2 +1,0 @@
-html,body{min-height:100%}body{background:#fff;color:#000;font-family:Arial, "Helvetica Neue", Helvetica, sans-serif;font-size:12px;margin:10px}
-/*# sourceMappingURL=style.css.map */
