--- www/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<title>Tensō</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-<link href="css/style.css" media="all" rel="stylesheet" type="text/css" />
-</head>
-
-<body>
-<h1>Tensō</h1>
-<p>Tensō is a REST framework for node.js, designed to simplify the implementation of APIs.</p>
-<p>You have successfully installed Tenso. Please create a `routes` module, `require()` it into your server config as "routes", & then restart the server.</p>
-
-<h2>Creating Routes</h2>
-<p>Routes are loaded as a module, with each HTTP method as an export, affording a very customizable API server.</p>
-<p>Route handlers have the context of the Tensō server, i.e. `this` will allow you to send a response with `this.respond( req, res, body[, status, headers] )`.</p>
-<p>The following example will create GET routes that will return an empty `Array` at `/`, an `Error` at `/reports/tps`, & a version 4 UUID at `/uuid`.</p>
-
-<pre>
-<code>
-var uuid = require( "keigai" ).util.uuid;
-
-module.exports.get = {
-  "/": [],
-  "/reports/tps": function ( req, res ) {
-    this.respond( req, res, new Error( "TPS Cover Sheet not attached" ), 785 );
-  },
-  "/uuid": function ( req, res ) {
-    this.respond( req, res, uuid() );
-  }
-}
-</code>
-</pre>
-</body>
-</html>
